implement api to query node doi, refactor doi generation code
