+++ desci-server/package.json
@@ -130,6 +130,7 @@
     "@types/multer-s3": "^3.0.1",
     "@types/node": "^20.10.4",
     "@types/supertest": "^2.0.12",
+    "@types/uuid": "^9.0.8",
     "@types/ws": "^8.5.10",
     "@types/xmldom": "^0.1.34",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
